asf_tcp_nf: Patch to correct typecast in tcp timestamp copy.
Earlier tcp timestamp is typecasted using long. But timestamp is of
4 byte. For 64-bit machine, long is 8 byte. This is incorrect
So it is typecasted using int.

+++ net/netfilter/nf_conntrack_proto_tcp.c
@@ -736,7 +736,7 @@ static bool tcp_in_window(const struct nf_conn *ct,
 					break;
 				case TCPOPT_TIMESTAMP:
 					state->seen[dir].td_tcptimestamp =
-						ntohl(*((unsigned long *)
+						ntohl(*((unsigned int *)
 							(tcpopt + 2)));
 					goto DONE;
 				default:
